requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
keyboard==0.13.5
pillow==10.4.0
requests==2.32.3
rnd-cat==0.0.4
urllib3==2.2.3
